Disable forced encryption when "guest" access is allowed
This took a long time to figure out, but it appears like Samba is unable
to create encrypted connections if the provided user does not have a
password associated with it (which is the case when the login falls back
to the "guest" user). Thus the connection fails with not so obvious errors
when encryption is enabled along with guest access.

The solution is to make shares with anonymous access much less secure, or
to create a "guest" user with a simple password. This change allows for
both of these approaches.